rename qa_dashboards -> dashboards
