+++ testdata/dashboards/qa_widgets.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from qa_dashboards.shared import (
+from dashboards.shared import (
     get_dashboard_desc_template,
     get_region_tmpl_var_presets,
 )
